AddCommentsandreworkpostpage

+++ src/components/Profil.js
@@ -5,11 +5,12 @@ import swal from "sweetalert";
 import AuthService from "../Middleware/authservice";
 import Footer from "../View/Footer";
 import logo from "../images/icongroupomanianoir.png";
-import EditProfile from "./editProfiles";
+import EditImage from "./editImage";
 const API_URL = 'http://localhost:5000/api/';
 
 const Register = () => {
-  
+  // On définit le state de notre composant Register pour stocker les données du formulaire
+  // we define the state of our Register component to store the data of the form
   const [nom, setNom] = useState("");
   const [prenom, setPrenom] = useState("");
   const [email, setEmail] = useState("");
@@ -18,32 +19,45 @@ const Register = () => {
   const history = useNavigate();
   let navigate = useNavigate();
   const user = AuthService.getCurrentUser();
-  const routeChange = () => {
-    let path = `/Connexion`;
-    navigate(path);
-  };
-  
+  // Suppression du Profil de l'utilisateur et redirection a la page de connexion
+  // Delete the profile of the user and redirect to the login page
   const deleteprofil = async (e) => {
     e.preventDefault();
-    try {
-      await axios.delete( API_URL + "user/", {
-        headers: {
-          Authorization: `Bearer ${user.token}`,
-        },
-      });
-
-      window.localStorage.clear();
-      alert("Votre compte a bien été supprimé");
-      navigate("/Connexion");
-    } catch (error) {
-      if (error.response) {
-        setMsg(error.response.data.msg);
+    //Utilisation de swal pour afficher un message de confirmation
+    // Use swal to display a confirmation message
+    swal({
+      title: "Etes-vous sûr ?",
+      text: "Vous voulez changer les informations de votre profil ?",
+      icon: "warning",
+      buttons: true,
+      dangerMode: true,
+    }).then((willDelete) => {
+      if (willDelete) {
+        try {
+          axios.delete( API_URL + "user/", {
+            headers: {
+              Authorization: `Bearer ${user.token}`,
+            },
+          });
+    
+          window.localStorage.clear();
+          swal("Votre profil a été supprimé avec succès", {
+            icon: "success",
+          });
+          navigate("/Connexion");
+        } catch (error) {
+          swal("Erreur lors de la modification du Profil", "", "error");
+        }
+      } else {
+        swal("La modification du profil a été annulée");
       }
-    }
+
+    });
   };
 
 
-
+    //Modification du profil de l'utilisateur et redirection au Dashboard
+    //Edit the profile of the user and redirect to the Dashboard
   const Editprofil = async (e) => {
     e.preventDefault();
     swal({
@@ -88,7 +102,8 @@ const Register = () => {
 
     });
   };
-
+    //Rendu du Profil de l'utilisateur
+    //Render the profile of the user
   
   return (
     <section className="hero has-background is-fullheight is-fullwidth has-background-grey-light ">
@@ -137,7 +152,9 @@ const Register = () => {
                   </button>
                 </div>
               </form>
-              <EditProfile/>
+              {/* Rendu du composant EditImage */}
+              {/* Rendering EditImage */}
+              <EditImage/>
                 
             </div>
             </div>
